+++ src/spdLevelGen/loopBuilder.ts
@@ -46,7 +46,8 @@ export class LoopBuilder extends RegularBuilder {
 	private randomBranchAngle = (r: Room): number => {
 		if (this.loopCenter === null) return SpdRandom.floatRange(0, 360);
 		const center = this.loopCenter;
-		let toCenter = angleBetweenPoints({ x: (r.left + r.right) / 2, y: (r.top + r.bottom) / 2 }, center);
+		// Java measures from a `float` PointF (`(left+right)/2f`) - narrow here.
+		let toCenter = angleBetweenPoints({ x: Math.fround((r.left + r.right) / 2), y: Math.fround((r.top + r.bottom) / 2) }, center);
 		if (toCenter < 0) toCenter = Math.fround(toCenter + 360);
 
 		let currAngle = SpdRandom.floatRange(0, 360);
